chore(vincent-tool-sdk): Move typescript type-checker files
